==> Geongnam_bigdata_analysis_competition/raw/p1_fetchOpenAPI.ipynb <==
! pip install pandas

!pip install requests

import pandas as pd
import requests
import json

# Define the parameters for the API request
parameters = {
    "key": "***",  # hide API key by using .gitignore
    "type": "json",       
    "numOfRows": 100,
    "pageNo": 5 # data consist of five pages in total
}

url = "https://data.ex.co.kr/openapi/basicinfo/unitList?"

# Send an HTTP GET request to the specified URL with the given parameters
response = requests.get(url, params=parameters)

# Print the HTTP status code (e.g., 200 for "OK", 404 for "Not Found")
# print(response.status_code)
# print(response.json())

response.json()['unitLists'][0]

# code = []
# name = []

for i in range(len(response.json()['unitLists'])):
    name.append(response.json()['unitLists'][i]['unitName'])
    code.append(response.json()['unitLists'][i]['unitCode'])

df = pd.DataFrame({'code':pd.Series(code), 'name':pd.Series(name)})
df

# save df in the csv format
df.to_csv('영업소코드.csv', index=False)
