==> Geongnam_bigdata_analysis_competition/raw/p2_checkRoute.ipynb <==
# install this module to rean a xlsx file
!pip install pandas openpyxl

# read an xlsx file
df2 = pd.read_excel('영업소간교통량240620.xlsx')
df2.head()

# create a new df from subset of the columns
df2_reduced = df2[['출발영업소코드','도착영업소코드','출발영업소명','도착영업소명','도착지방향총교통량','출발지방향총교통량']]
df2_reduced.head()

startCode = 243
destCode = 241
bMask1 = (df2_reduced['출발영업소코드']==startCode) & (df2_reduced['도착영업소코드']==destCode)
# bMask2 = (df2_reduced['출발영업소코드']==destCode) & (df2_reduced['도착영업소코드']==startCode)
print(df2_reduced[bMask1])
# print(df2_reduced[bMask2])

code = 239
bMask3 = (df2_reduced['출발영업소코드']==code) | (df2_reduced['도착영업소코드']==code)
print(df2_reduced[bMask3])
