==> Geongnam_bigdata_analysis_competition/raw/p3_avgFourDays.ipynb <==
# average figures measured on June 1st(Sat), 4th(Tue), 16th(Sun) and 20th(Thu)
# import four xlsx files
import pandas as pd
df01 = pd.read_excel('영업소간교통량240601.xlsx')
df04 = pd.read_excel('영업소간교통량240604.xlsx')
df16 = pd.read_excel('영업소간교통량240616.xlsx')
df20 = pd.read_excel('영업소간교통량240620.xlsx')

# extract columns of interest
# pdf stands for Partial dataFrame
pdf01 = df01[['출발영업소코드','도착영업소코드','출발영업소명','도착영업소명','도착지방향총교통량','출발지방향총교통량']]
pdf04 = df04[['출발영업소코드','도착영업소코드','출발영업소명','도착영업소명','도착지방향총교통량','출발지방향총교통량']]
pdf16 = df16[['출발영업소코드','도착영업소코드','출발영업소명','도착영업소명','도착지방향총교통량','출발지방향총교통량']]
pdf20 = df20[['출발영업소코드','도착영업소코드','출발영업소명','도착영업소명','도착지방향총교통량','출발지방향총교통량']]

# sort pdfs by columns '출발영업소코드' and'도착영업소코드'
pdf01 = pdf01.sort_values(by=['출발영업소코드','도착영업소코드'], ascending=[True,True], ignore_index=True)
pdf04 = pdf04.sort_values(by=['출발영업소코드','도착영업소코드'], ascending=[True,True], ignore_index=True)
pdf16 = pdf16.sort_values(by=['출발영업소코드','도착영업소코드'], ascending=[True,True], ignore_index=True)
pdf20 = pdf20.sort_values(by=['출발영업소코드','도착영업소코드'], ascending=[True,True], ignore_index=True)

pdf01 # rudundant rows are found at its tail

# truncate rows with invalide station codes by boolean indexing
# 987 is the maximum code according to the list obtained in the first step(fetchOpentAPI_1)
# tdf stands for truncated dataFrame
codeLimit = 987
bMask01 = (pdf01['출발영업소코드']>codeLimit) | (pdf01['도착영업소코드']>codeLimit)
tdf01 = pdf01[~bMask01]
print("tdf01 :\n", tdf01)

# do the same to the rest
# as all are thought to have the same rows, use the bMask01 to the rest as well
tdf04 = pdf04[~bMask01]
tdf16 = pdf16[~bMask01]
tdf20 = pdf20[~bMask01]

# check their tails
print("tdf04 :\n", tdf04.tail())
print("tdf16 :\n", tdf16.tail())
print("tdf20 :\n", tdf20.tail())

# make copies of tdfs and rename traffic columns
mdf01 = tdf01.copy()
mdf01 = mdf01.rename(columns={'도착지방향총교통량':'도착지방향총교통량01','출발지방향총교통량':'출발지방향총교통량01'})

mdf04 = tdf04.copy()
mdf04 = mdf04.rename(columns={'도착지방향총교통량':'도착지방향총교통량04','출발지방향총교통량':'출발지방향총교통량04'})

mdf16 = tdf16.copy()
mdf16 = mdf16.rename(columns={'도착지방향총교통량':'도착지방향총교통량16','출발지방향총교통량':'출발지방향총교통량16'})

mdf20 = tdf20.copy()
mdf20 = mdf20.rename(columns={'도착지방향총교통량':'도착지방향총교통량20','출발지방향총교통량':'출발지방향총교통량20'})

# merge mdfs one by one
mdf = pd.merge(mdf01, mdf04, how='outer')
mdf = pd.merge(mdf, mdf16, how='outer')
mdf = pd.merge(mdf, mdf20, how='outer')
mdf

# add average columns
mdf['도착지방향평균'] = (mdf['도착지방향총교통량01'] + mdf['도착지방향총교통량04'] + mdf['도착지방향총교통량16'] + mdf['도착지방향총교통량20'])/4
mdf['출발지방향평균'] = (mdf['출발지방향총교통량01'] + mdf['출발지방향총교통량04'] + mdf['출발지방향총교통량16'] + mdf['출발지방향총교통량20'])/4
mdf

# finally get the desired dataFrame
# fdf stands for final dataFrame
fdf = mdf[['출발영업소코드','도착영업소코드','출발영업소명','도착영업소명','도착지방향평균','출발지방향평균']]
fdf

# save fdf for later use
fdf.to_csv('fdf.csv', index=False)
