==> Geongnam_bigdata_analysis_competition/raw/p4_prepareDfs.ipynb <==
# as it turned out that the data sets were incompatible...
# employ another data file with modifications based on wikiPage concerning highway stations :
# (https://ko.wikipedia.org/wiki/%EB%8C%80%ED%95%9C%EB%AF%BC%EA%B5%AD%EC%9D%98_%EA%B3%A0%EC%86%8D%EB%8F%84%EB%A1%9C_%EC%98%81%EC%97%85%EC%86%8C_%EB%AA%A9%EB%A1%9D)
import pandas as pd

# import fdf from the previous work 'avgFourDays_3'
fdf = pd.read_csv('fdf.csv')
print("fdf :\n", fdf.head())

# edit out redundant columns
fdf = fdf.drop(columns=['출발영업소코드','도착영업소코드','출발지방향평균'])
print("fdf :\n", fdf.head())

# this is our brand new sdf
sdf = pd.read_excel('새영업소현황.xlsx')
print("sdf :\n", sdf.keys())
print("\n", sdf.head())

sdf = sdf.drop(columns=['전화번호','팩스번호','우편번호\n(5자리)'])
sdf.head()
# in the new xlsx sheet, 새영업소명 is added to match names of stations to those of fdf

# as the old column 영업소명 is replaced by 새명업소명, let's drop it
sdf = sdf.drop(['영업소명'], axis=1)
sdf.head()

# just like in the previous work, add new addresses 세부주소1 and 세부주소2
sdf['세부주소1'] = sdf['주소'].apply(lambda x: x.split()[0]) 
sdf['세부주소2'] = sdf['주소'].apply(lambda x: x.split()[0] if x.split()[0]!='경상남도' else x.split()[1])
sdf.head()

# get lists of names ; in- and outside 경상남도
others = sdf['세부주소1'].unique()
others

import numpy as np
# np.delete(arrayName, index)
others = sdf['세부주소1'].unique()
for idx, val in enumerate(others):
    if val in ['예정','경상남도']:
        others = np.delete(others, idx)
others # why need to execute this cell twice ?

# alternatively.
others = sdf['세부주소1'].unique()
for elem in ['예정','경상남도']:
    others = np.delete(others, np.where(others==elem))
others

gyeongnam = sdf['세부주소2'].unique()
for elem in others:
    gyeongnam = np.delete(gyeongnam, np.where(gyeongnam==elem))
gyeongnam

# drop the column of complete addresses
sdf = sdf.drop(['주소'], axis=1)
sdf.head()

# save for later use
sdf.to_csv('sdf.csv', index=False) 

# join the two dfs, fdf and sdf
# first to add address columns for start-stations
# mdf stands for merged dataFrame
mdf = pd.merge(fdf, sdf, left_on="출발영업소명", right_on="새영업소명").drop('새영업소명', axis=1)
mdf = pd.merge(mdf, sdf, left_on="도착영업소명", right_on="새영업소명").drop('새영업소명', axis=1)
mdf.head()

# drop redundant columns
mdf = mdf.drop(columns=['노선명_x','노선명_y'])
mdf.head()

# rename the added columns
mdf = mdf.rename(columns={'세부주소1_x':'출발주소1', '세부주소2_x':'출발주소2',
                         '세부주소1_y':'도착주소1', '세부주소2_y':'도착주소2'})
mdf.head()

# change the order of columns
mdf = mdf.iloc[:,[0,3,4,1,5,6,2]]
mdf.head()

# drop rows having nothing to do with 경상남도 by using 도착주소1 which refers to province
# namely, leave no rows whose start- and terminal stations are both outside 경상남도
dropCond = mdf[(mdf['출발주소1']!='경상남도') & (mdf['도착주소1']!='경상남도')].index
mdf = mdf.drop(dropCond)
mdf.head()

## add lately
# for the network over all provinces, take some parts of mdf
# the first partial mdf
pmdf = mdf[['출발주소1','도착주소1','도착지방향평균']]
pmdf.to_csv('전국교통량평균.csv', index=False, encoding='utf-8') 

## not excecuted accordingly

# then, first addresses are redundant
# a corresponding adjacency matrix will be constructed based on the second addresses
mdf = mdf.drop(columns=['출발주소1','도착주소1'])

# rename the second ones
mdf = mdf.rename(columns={'출발주소2':'출발영업소주소', '도착주소2':'도착영업소주소'})
mdf.head()

# save this dataFrame in csv format
mdf.to_csv('mdf.csv', index=False) 

# end of the code
