==> Geongnam_bigdata_analysis_competition/raw/p4_trial_prepareDfs.ipynb <==
import pandas as pd

# import fdf from the previous work 'avgFourDays_3'
fdf = pd.read_csv('fdf.csv')
print("fdf :\n", fdf.head())

# sdf stands for station dataFrame
sdf = pd.read_excel('노선별영업소현황.xlsx')
print("sdf :\n", sdf.head())

# feature engineering on sdf['주소'] before it is joined with fdf
# with the use of .split() method. The followings are examples :
print(sdf['주소'][0])
print(sdf['주소'][0].split())
print(sdf['주소'][0].split()[0])
print(sdf['주소'][0].split()[1])

# 세부주소1 refers to province, while 세부주소2 does to province or municipality
# the latter depends on whether a region concerns 경상남도
sdf['세부주소1'] = sdf['주소'].apply(lambda x: x.split()[0]) 
sdf['세부주소2'] = sdf['주소'].apply(lambda x: x.split()[0] if x.split()[0]!='경상남도' else x.split()[1])

# check the following two cells out :
# as for stations outside 경상남도, both detailed addresses have the same value refering to province
# for example,
bMask = (sdf['영업소명']=='북다사') | (sdf['영업소명']=='양양')
sdf[bMask]

# on contrary, 세부주소2 refers to a certain municipality when it comes to stations inside 경상남도
bMask = (sdf['영업소명']=='양산') | (sdf['영업소명']=='장유')
sdf[bMask]

# for the sake of simplicity, drop redundant columns
# 노선명 won't be directly used, either, but let's just leave it there for reference
sdf = sdf.drop(columns=['전화번호','팩스번호','우편번호\n(5자리)','주소'])
sdf.head()

# before join the two fdf and sdf on 영업소명, check whether the two 영업소명 are compatible
fdfStation1 = sorted(fdf['출발영업소명'].unique())
fdfStation2 = sorted(fdf['도착영업소명'].unique())
sdfStation = sorted(sdf['영업소명'].unique())

# check the number of their elements
print(len(fdfStation1))
print(len(fdfStation2))
print(len(sdfStation))

# well they do not match well !

# construct a dataFrame with them and convert it to csv
compStation = pd.DataFrame({'fdf':pd.Series(fdfStation1), 'sdf':pd.Series(sdfStation)})
compStation.to_csv('compStation.csv', index=False)

# the data list is quite long!
# better filter out lines passing through 경상남도


myNode = ['거제시','거창군','고성군','김해시','남해군','밀양시','사천시','산청군','양산시','의령군','진주시','창녕군'
         ,'창원시','통영시','하동군','함안군','함양군','합천군','경상북도','전라북도','전라남도','울산광역시','부산광역시']
len(myNode)

import numpy as np
adjMatrix = pd.DataFrame(np.zeros((23,23)), columns=myNode, index=myNode)
adjMatrix.head()

# use the code from the previous work 'checkRoute_2' for reference
# when it comes to 경부선(부산->양산)
## check the figures again before adding them to the adjMatrix 
startCode = 529
destCode = 138
bMask1 = (df['출발영업소코드']==startCode) & (df['도착영업소코드']==destCode)
bMask2 = (df['출발영업소코드']==destCode) & (df['도착영업소코드']==startCode)
print(df[bMask1])
print(df[bMask2])

# fill up adjMatrix with data from df
# forward
adjMatrix['부산광역시']['양산시'] += df[bMask1]['도착지방향평균']
# backward
adjMatrix['양산시']['부산광역시'] += df[bMask1]['도착지방향평균']
