==> Geongnam_bigdata_analysis_competition/raw/p5_constructAdjMatrix.ipynb <==
# construct an adjacency matrix with the use of mdf created in the previous work 'prepareDfs_4'
import pandas as pd
mdf = pd.read_csv('mdf.csv')
print("mdf :\n", mdf.head())

# get a node list from mdf['출발영업소주소']
nList = mdf['출발영업소주소'].unique()
print(len(nList))
print(nList)

# the last element concerns '두물머리' station which has not established yet
# figures concerning this will be removed after filling up the matrix first

# create an empty matrix to fill up
adjMatrix = pd.DataFrame(np.zeros((len(nList),len(nList))), columns=nList, index=nList)
adjMatrix.head()

# while iterating over the rows, fill up the elements
for index, row in mdf.iterrows():
    adjMatrix[row['출발영업소주소']][row['도착영업소주소']] += row['도착지방향평균']

adjMatrix.head()

# check out '예정'; sure there is no data related with a station not running right now!
adjMatrix['예정']

# drop concerning row and column, respectively
adjMatrix = adjMatrix.drop(['예정'], axis=0)
adjMatrix = adjMatrix.drop(['예정'], axis=1)

# save the adjMatrix in csv format for later use
adjMatrix.to_csv('adjMatrix.csv', index=True, encoding='utf-8') 

# end of code
