==> Geongnam_bigdata_analysis_competition/raw/p9_transTime.ipynb <==
import pandas as pd
import numpy as np

station = pd.read_excel('ETC영업소주소.xlsx') # addresses added to ETC_영업소주소.xlsx
station.head()

# the API takes too long to load, use xlsx files instead
# data obtained at 7am (the busiest hour) and concerns suv only
# namely, 집계시=7 & TCS차종구분코드=1
trans0604 = pd.read_excel('영업소간통행시간240604_reduced.xlsx')
trans0604 = trans0604.sort_values(by=['출발영업소코드','도착영업소코드'],ignore_index=True)
trans0604 = trans0604.rename(columns={'통행시간':'통행시간04'})
trans0604.head()

# load the others
trans0612 = pd.read_excel('영업소간통행시간240612_reduced.xlsx')
trans0612 = trans0612.sort_values(by=['출발영업소코드','도착영업소코드'],ignore_index=True)
trans0612 = trans0612.rename(columns={'통행시간':'통행시간12'})
trans0612.head()

trans0620 = pd.read_excel('영업소간통행시간240620_reduced.xlsx')
trans0620 = trans0620.sort_values(by=['출발영업소코드','도착영업소코드'],ignore_index=True)
trans0620 = trans0620.rename(columns={'통행시간':'통행시간20'})
trans0620.head()

trans0624 = pd.read_excel('영업소간통행시간240624_reduced.xlsx')
trans0624 = trans0624.sort_values(by=['출발영업소코드','도착영업소코드'],ignore_index=True)
trans0624 = trans0624.rename(columns={'통행시간':'통행시간24'})
trans0624.tail()

# create a new df containing the list of avg values
trans = trans0604.copy()
trans = pd.merge(trans, trans0612, how='outer')
trans = pd.merge(trans, trans0620, how='outer')
trans = pd.merge(trans, trans0624, how='outer')
trans

# drop redundant columns
trans = trans.drop(columns=['집계시','TCS차종구분코드'])
trans

# take the average of the four transition time with NaNs ignored
tempZip = list(zip(trans['통행시간04'].values,trans['통행시간12'].values,
           trans['통행시간20'].values,trans['통행시간24'].values))
avgTrans = []
for i in range(len(tempZip)):
    avgTrans.append(np.nanmean(tempZip[i]))
trans['통행시간평균'] = pd.Series(avgTrans)
trans.head()

trans = trans.drop(columns=['통행시간04','통행시간12','통행시간20','통행시간24'])
trans.head()

# merge the two dfs with respect to station code
mdf = pd.merge(trans, station, left_on='출발영업소코드', right_on='영업소코드').drop('영업소코드', axis=1)
mdf = pd.merge(mdf, station, left_on='도착영업소코드', right_on='영업소코드').drop('영업소코드', axis=1)
mdf.head()

# take the same steps as in p4 to get an associate adjacency matrix
mdf = mdf.rename(columns={'세부주소1_x':'출발주소1', '세부주소2_x':'출발주소2',
                         '세부주소1_y':'도착주소1', '세부주소2_y':'도착주소2',
                         '영업소명_x':'출발영업소','영업소명_y':'도착영업소'})
mdf.head()

mdf = mdf.drop(columns=['출발영업소코드','도착영업소코드'])
mdf.head()

mdf = mdf.iloc[:,[1,2,3,4,5,6,0]]
mdf.head()

# truncate the average values
import math
mdf['통행시간평균'] = mdf['통행시간평균'].map(lambda x:math.trunc(x))
mdf.head()

# the first checkpoint - save mdf
mdf.to_csv('영업소간통행시간평균.csv')

# drop rows having nothing to do with 경상남도 by using 도착주소1 which refers to province
# namely, leave no rows whose start- and terminal stations are both outside 경상남도
dropCond = mdf[(mdf['출발주소1']!='경상남도') & (mdf['도착주소1']!='경상남도')].index
mdf = mdf.drop(dropCond)
mdf.head()

# then, first addresses are redundant
# a corresponding adjacency matrix will be constructed based on the second addresses
mdf = mdf.drop(columns=['출발주소1','도착주소1'])

# rename the second ones
mdf = mdf.rename(columns={'출발주소2':'출발영업소주소', '도착주소2':'도착영업소주소'})
mdf.head()

# the second checkpoint - save mdf
mdf.to_csv('경남영업소간통행시간평균.csv', index=False) 

# end of code
