==> Geongnam_bigdata_analysis_competition/raw/p10_transTAdjMatrix.ipynb <==
# construct an adjacency matrix with the use of mdf created in the previous work p9
import pandas as pd
mdf = pd.read_csv('경남영업소간통행시간평균.csv')
mdf.head()

# get a node list from mdf['도착영업소주소']
nList = mdf['도착영업소주소'].unique()
print(len(nList))
print(nList)

# create an empty matrix to fill up
import numpy as np
adjMatrix = pd.DataFrame(np.zeros((len(nList),len(nList))), columns=nList, index=nList)
adjMatrix.head()

# while iterating over the rows, fill up the elements
for index, row in mdf.iterrows():
    adjMatrix[row['출발영업소주소']][row['도착영업소주소']] += row['통행시간평균']
adjMatrix.head()

# save
adjMatrix.to_csv('transTimeAdjMatrix.csv', index=True, encoding='utf-8') 

# nullify its diagonal elements
for i in range(len(nList)):
    adjMatrix.iloc[i][i] = 0
adjMatrix.head()

# and save it
adjMatrix.to_csv('transTimeAdjMatrixZeroDiagonalElem.csv', index=True, encoding='utf-8') 

# the matrix is asymmetric 
# end of the code
