==> Geongnam_bigdata_analysis_competition/raw/p12_weekdaysTotalTraffic.ipynb <==
# get a new data set concerning four weekdays - June 4, 12, 20, and 24
import pandas as pd
df04 = pd.read_excel('영업소간교통량240604.xlsx')

df12 = pd.read_excel('영업소간교통량240612.xlsx')
df20 = pd.read_excel('영업소간교통량240620.xlsx')
df24 = pd.read_excel('영업소간교통량240624.xlsx')

# extract columns of interest
# pdf stands for Partial dataFrame
pdf04 = df04[['출발영업소코드','도착영업소코드','출발영업소명','도착영업소명','도착지방향총교통량','출발지방향총교통량']]
pdf12 = df12[['출발영업소코드','도착영업소코드','출발영업소명','도착영업소명','도착지방향총교통량','출발지방향총교통량']]
pdf20 = df20[['출발영업소코드','도착영업소코드','출발영업소명','도착영업소명','도착지방향총교통량','출발지방향총교통량']]
pdf24 = df24[['출발영업소코드','도착영업소코드','출발영업소명','도착영업소명','도착지방향총교통량','출발지방향총교통량']]

# sort pdfs by columns '출발영업소코드' and'도착영업소코드'
pdf04 = pdf04.sort_values(by=['출발영업소코드','도착영업소코드'], ascending=[True,True], ignore_index=True)
pdf12 = pdf12.sort_values(by=['출발영업소코드','도착영업소코드'], ascending=[True,True], ignore_index=True)
pdf20 = pdf20.sort_values(by=['출발영업소코드','도착영업소코드'], ascending=[True,True], ignore_index=True)
pdf24 = pdf24.sort_values(by=['출발영업소코드','도착영업소코드'], ascending=[True,True], ignore_index=True)

pdf04 # rudundant rows are found at its tail

# truncate rows with invalide station codes by boolean indexing
# 987 is the maximum code according to the list obtained in the first step(fetchOpentAPI_1)
# tdf stands for truncated dataFrame
codeLimit = 987
bMask = (pdf04['출발영업소코드']>codeLimit) | (pdf04['도착영업소코드']>codeLimit)
tdf04 = pdf04[~bMask]
print("tdf01 :\n", tdf04)

# do the same to the rest
# as all are thought to have the same rows, use the bMask01 to the rest as well
tdf12 = pdf12[~bMask]
tdf20 = pdf20[~bMask]
tdf24 = pdf24[~bMask]

# check their tails
print("tdf04 :\n", tdf12.tail())
print("tdf16 :\n", tdf20.tail())
print("tdf20 :\n", tdf24.tail())

# make copies of tdfs and rename traffic columns
mdf04 = tdf04.copy()
mdf04 = mdf04.rename(columns={'도착지방향총교통량':'도착지방향총교통량04','출발지방향총교통량':'출발지방향총교통량04'})

mdf12 = tdf12.copy()
mdf12 = mdf12.rename(columns={'도착지방향총교통량':'도착지방향총교통량12','출발지방향총교통량':'출발지방향총교통량12'})

mdf20 = tdf20.copy()
mdf20 = mdf20.rename(columns={'도착지방향총교통량':'도착지방향총교통량20','출발지방향총교통량':'출발지방향총교통량20'})

mdf24 = tdf24.copy()
mdf24 = mdf24.rename(columns={'도착지방향총교통량':'도착지방향총교통량24','출발지방향총교통량':'출발지방향총교통량24'})

mdf = pd.merge(mdf04, mdf12, how='outer')
mdf = pd.merge(mdf, mdf20, how='outer')
mdf = pd.merge(mdf, mdf24, how='outer')
mdf

# add average columns
mdf['도착지방향평균'] = (mdf['도착지방향총교통량04'] + mdf['도착지방향총교통량12'] + mdf['도착지방향총교통량20'] + mdf['도착지방향총교통량24'])/4
mdf['출발지방향평균'] = (mdf['출발지방향총교통량04'] + mdf['출발지방향총교통량12'] + mdf['출발지방향총교통량20'] + mdf['출발지방향총교통량24'])/4
mdf

# finally get the desired dataFrame
# fdf stands for final dataFrame
fdf = mdf[['출발영업소코드','도착영업소코드','출발영업소명','도착영업소명','도착지방향평균','출발지방향평균']]
fdf

# save fdf for later use
fdf.to_csv('평일전국교통량평균.csv', index=False) 

# edit out redundant columns
# fdf = fdf.drop(columns=['출발영업소코드','도착영업소코드','출발지방향평균']) excecuted
fdf.head()

# import data concerning location of stations - sdf
sdf = pd.read_csv('sdf.csv',encoding='utf-8')
sdf.head()

# join the two dfs, fdf and sdf
# first to add address columns for start-stations
# mdf stands for merged dataFrame
mdf = pd.merge(fdf, sdf, left_on="출발영업소명", right_on="새영업소명").drop('새영업소명', axis=1)
mdf = pd.merge(mdf, sdf, left_on="도착영업소명", right_on="새영업소명").drop('새영업소명', axis=1)
mdf.head()

# drop redundant columns
mdf = mdf.drop(columns=['노선명_x','노선명_y'])
mdf.head()

# rename the added columns
mdf = mdf.rename(columns={'세부주소1_x':'출발주소1', '세부주소2_x':'출발주소2',
                         '세부주소1_y':'도착주소1', '세부주소2_y':'도착주소2'})
mdf.head()

# change the order of columns
mdf = mdf.iloc[:,[0,3,4,1,5,6,2]]
mdf.head()

# for the network over all provinces, take some parts of mdf
# the first partial mdf
pmdf = mdf[['출발주소1','도착주소1','도착지방향평균']]
pmdf.to_csv('전국교통량평균.csv', index=False, encoding='utf-8') 

import pandas as pd
pmdf = pd.read_csv('전국교통량평균.csv')
pmdf.head()

# adjMatrix associated with pmdf
import numpy as np
nList = np.concatenate((pmdf['출발주소1'].values, pmdf['도착주소1'].values), axis=None)
nList = np.unique(nList)

# create an empty matrix to fill up
import numpy as np
adjMatrix = pd.DataFrame(np.zeros((len(nList),len(nList))), columns=nList, index=nList)
adjMatrix.head()

# while iterating over the rows, fill up the elements
for index, row in pmdf.iterrows():
    adjMatrix[row['출발주소1']][row['도착주소1']] += row['도착지방향평균']
    adjMatrix[row['도착주소1']][row['출발주소1']] += row['도착지방향평균']
adjMatrix.head()

# clear self-loops
for i in range(len(nList)):
    adjMatrix.iloc[i][i] = 0
adjMatrix.head()

adjMatrix = adjMatrix.drop(['예정'], axis=0)
adjMatrix = adjMatrix.drop(['예정'], axis=1)

# total traffic
totalTraff = adjMatrix.sum(axis='columns')
totalTraff = totalTraff.apply(lambda x:int(x))

# save the adjMatrix in csv format for later use
adjMatrix.to_csv('평일전국교통량행렬.csv', index=True, encoding='utf-8') 

totalTraff.to_csv('평일전국총교통량.csv')

# end of code
