==> Geongnam_bigdata_analysis_competition/raw/p17_checkWeather.ipynb <==
# get the open API '휴게소별 날씨 정보' provided by 한국도로공사
import pandas as pd
import requests
import json

# Define the parameters for the API request
parameters = {
    "key": "***",  # hide API key by using .gitignore
    "type": "json",       
    "sdate": "20240624", # check dates - 4th, 12th, 20th, 24th
    "stdHour": "10" # weather data seem to be acquired from 10am..
}
url = "https://data.ex.co.kr/openapi/restinfo/restWeatherList"
response = requests.get(url, params=parameters)

response.json()['list'][0]

stName = []
address = []
weather = []
dew = []
rainfall = []

for i in range(len(response.json()['list'])):
    stName.append(response.json()['list'][i]['unitName'])
    address.append(response.json()['list'][i]['addr'])
    weather.append(response.json()['list'][i]['weatherContents'])
    dew.append(response.json()['list'][i]['dewValue'])
    rainfall.append(response.json()['list'][i]['rainfallValue'])

df = pd.DataFrame({'stName':pd.Series(stName), 'address':pd.Series(address), 'weather':pd.Series(weather),
                  'dew':pd.Series(dew), 'rain':pd.Series(rainfall)})
df.head()

df['dew'].unique()

df['rain'].unique()

df['dew'] = df['dew'].apply(lambda x:int(float(x)))
df['rain'] = df['rain'].apply(lambda x:int(float(x)))

# drop rows whose value is just set as blank
dropIdx = df[(df['address'] =='')|(df['rain']==-99)].index
df = df.drop(dropIdx)
df.head()

# no rain at any stations

## previous trial

df['newAdd'] = df['address'].apply(lambda x:x.split()[0])
df.head()

df = df.drop('address',axis=1)
df.head()

df['weather'].unique()

df[df['weather']=='구름많음']
