==> Geongnam_bigdata_analysis_competition/raw/p18_busyhourTraffic.ipynb <==
import pandas as pd
df1 = pd.read_excel('이용차량서울경기.xlsx')
df1.head()

df2 = pd.read_excel('이용차량강원.xlsx')
df3 = pd.read_excel('이용차량대전충남.xlsx')
df4 = pd.read_excel('이용차량충북.xlsx')
df5 = pd.read_excel('이용차량대구경북.xlsx')
df6 = pd.read_excel('이용차량전북.xlsx')
df7 = pd.read_excel('이용차량광주전남.xlsx')
df8 = pd.read_excel('이용차량부산경남.xlsx')

# date column needs splitting
df1['일자'] = df1['일자'].apply(lambda x:x.split('.')[2])
df1.head()

df2['일자'] = df2['일자'].apply(lambda x:x.split('.')[2])
df3['일자'] = df3['일자'].apply(lambda x:x.split('.')[2])
df4['일자'] = df4['일자'].apply(lambda x:x.split('.')[2])
df5['일자'] = df5['일자'].apply(lambda x:x.split('.')[2])
df6['일자'] = df6['일자'].apply(lambda x:x.split('.')[2])
df7['일자'] = df7['일자'].apply(lambda x:x.split('.')[2])
df8['일자'] = df8['일자'].apply(lambda x:x.split('.')[2])

df8.head()

getRow = [i for j in [4, 12, 20, 24] for i in range(3*(j-1), 3*(j-1) + 3)]
getRow

getRow = [i for j in [4, 12, 20, 24] for i in range(3*(j-1), 3*(j-1) + 3)]
print(getRow)
wdf1 = df1.loc[getRow]
wdf1

wdf2 = df2.loc[getRow]
wdf3 = df3.loc[getRow]
wdf4 = df4.loc[getRow]
wdf5 = df5.loc[getRow]
wdf6 = df6.loc[getRow]
wdf7 = df7.loc[getRow]
wdf8 = df8.loc[getRow]

# filter out necessary columns
wdf1 = wdf1[['일자','입출구','08시']]
wdf1

# wdf1 = wdf1.drop('일자',axis=1) excuted
s1 = wdf1.groupby('입출구').mean()
s1

wdf2 = wdf2.drop('일자',axis=1)
wdf3 = wdf3.drop('일자',axis=1)
wdf4 = wdf4.drop('일자',axis=1)
wdf5 = wdf5.drop('일자',axis=1)
wdf6 = wdf6.drop('일자',axis=1)
wdf7 = wdf7.drop('일자',axis=1)
wdf8 = wdf8.drop('일자',axis=1)

s2 = wdf2.groupby('입출구').mean()
s3 = wdf3.groupby('입출구').mean()
s4 = wdf4.groupby('입출구').mean()
s5 = wdf5.groupby('입출구').mean()
s6 = wdf6.groupby('입출구').mean()
s7 = wdf7.groupby('입출구').mean()
s8 = wdf8.groupby('입출구').mean()

s1

s1.loc['입구'].values[0]

entList = []
exitList = []

entList.append(s1.loc['입구'].values[0])
entList.append(s2.loc['입구'].values[0])
entList.append(s3.loc['입구'].values[0])
entList.append(s4.loc['입구'].values[0])
entList.append(s5.loc['입구'].values[0])
entList.append(s6.loc['입구'].values[0])
entList.append(s7.loc['입구'].values[0])
entList.append(s8.loc['입구'].values[0])

exitList.append(s1.loc['출구'].values[0])
exitList.append(s2.loc['출구'].values[0])
exitList.append(s3.loc['출구'].values[0])
exitList.append(s4.loc['출구'].values[0])
exitList.append(s5.loc['출구'].values[0])
exitList.append(s6.loc['출구'].values[0])
exitList.append(s7.loc['출구'].values[0])
exitList.append(s8.loc['출구'].values[0])

entList

# load ent- and exit traffic ration adjacent matrix
exitTraff = pd.read_csv('newExitRatioAdjMatrix.csv')
exitTraff

exitTraff = exitTraff.drop('exitTotalTraff', axis=1)

exitTraff # needs its upper triangle

entTraff = pd.read_csv('newEntRatioAdjMatrix.csv')

entTraff.drop(entTraff.tail(1).index, inplace=True)

entTraff # needs its lower triangle

entTraff['name'].values

len(entTraff['name'])

entTraff.iloc[0] # row idex starts with zero

entTraff.iloc[0][2] # yet column index does with one

# create an empty df and fill it up with the two
traffCombined = pd.DataFrame({'서울경기':np.zeros(8),'강원도':np.zeros(8),'대전충남':np.zeros(8),'충청북도':np.zeros(8),
'대구경북':np.zeros(8),'전라북도':np.zeros(8),'광주전남':np.zeros(8),'부산경남':np.zeros(8)}, index=entTraff['name'].values)
traffCombined

# fill its upper triangle
for i in range(len(entTraff['name'])):
    for j in range(i+1,len(entTraff['name'])):
        traffCombined.iloc[i][j] = entTraff.iloc[i][j+1]
traffCombined

# fill its lower triangle
for i in range(1,len(entTraff['name'])):
    for j in range(i):
        traffCombined.iloc[i][j] = exitTraff.iloc[i][j+1]
traffCombined

# mutiply 입출구통행량
# fill its upper triangle conerning exit
for i in range(len(entTraff['name'])):
    for j in range(i+1,len(entTraff['name'])):
        traffCombined.iloc[i][j] = traffCombined.iloc[i][j] * entList[i]
traffCombined

# fill its lower triangle conerning entrance
for i in range(1,len(entTraff['name'])):
    for j in range(i):
        traffCombined.iloc[i][j] = traffCombined.iloc[i][j] * exitList[i]
traffCombined

traffCombined.to_csv('predict8amTraffic.csv', encoding='utf-8')
