==> Geongnam_bigdata_analysis_competition/raw/p32_closeLook.ipynb <==
import pandas as pd
import numpy as np
import matplotlib as mpl
import matplotlib.pyplot as plt
import networkx as nx
import matplotlib.font_manager as fm
fm.fontManager.addfont('BinggraeII.ttf')

df = pd.read_excel('provinceTraffic_reduced&trimed.xlsx',sheet_name='sorted')
df

# single out routes
pdf = df.iloc[[0,1,7,8,13,47,48,49],:]
pdf = pdf.reset_index(drop=True)
pdf

# import population data
pop = pd.read_excel('24년2월경남인구.xlsx')
pop

ppop = pop.iloc[[0,4,5,7],:]
ppop = ppop.reset_index(drop=True)
ppop

ppop.loc[4] = ['부산광역시',3311402]
ppop
